+++ projects/clr/hipamd/include/hip/amd_detail/hip_prof_str.h
@@ -509,7 +509,8 @@ enum hip_api_id_t {
   HIP_API_ID_hipMemGetDefaultMemPool = 484,
   HIP_API_ID_hipDeviceGetLuid = 485,
   HIP_API_ID_hipInitDevice = 486,
-  HIP_API_ID_LAST = 486,
+  HIP_API_ID_hipDeviceFlushGPUDirectRDMAWrites = 487,
+  HIP_API_ID_LAST = 487,
 
   HIP_API_ID_hipChooseDevice = HIP_API_ID_CONCAT(HIP_API_ID_,hipChooseDevice),
   HIP_API_ID_hipGetDeviceProperties = HIP_API_ID_CONCAT(HIP_API_ID_,hipGetDeviceProperties),
@@ -585,6 +586,7 @@ static inline const char* hip_api_name(const uint32_t id) {
     case HIP_API_ID_hipDeviceComputeCapability: return "hipDeviceComputeCapability";
     case HIP_API_ID_hipDeviceDisablePeerAccess: return "hipDeviceDisablePeerAccess";
     case HIP_API_ID_hipDeviceEnablePeerAccess: return "hipDeviceEnablePeerAccess";
+    case HIP_API_ID_hipDeviceFlushGPUDirectRDMAWrites: return "hipDeviceFlushGPUDirectRDMAWrites";
     case HIP_API_ID_hipDeviceGet: return "hipDeviceGet";
     case HIP_API_ID_hipDeviceGetAttribute: return "hipDeviceGetAttribute";
     case HIP_API_ID_hipDeviceGetByPCIBusId: return "hipDeviceGetByPCIBusId";
@@ -1065,6 +1067,7 @@ static inline uint32_t hipApiIdByName(const char* name) {
   if (strcmp("hipDeviceComputeCapability", name) == 0) return HIP_API_ID_hipDeviceComputeCapability;
   if (strcmp("hipDeviceDisablePeerAccess", name) == 0) return HIP_API_ID_hipDeviceDisablePeerAccess;
   if (strcmp("hipDeviceEnablePeerAccess", name) == 0) return HIP_API_ID_hipDeviceEnablePeerAccess;
+  if (strcmp("hipDeviceFlushGPUDirectRDMAWrites", name) == 0) return HIP_API_ID_hipDeviceFlushGPUDirectRDMAWrites;
   if (strcmp("hipDeviceGet", name) == 0) return HIP_API_ID_hipDeviceGet;
   if (strcmp("hipDeviceGetAttribute", name) == 0) return HIP_API_ID_hipDeviceGetAttribute;
   if (strcmp("hipDeviceGetByPCIBusId", name) == 0) return HIP_API_ID_hipDeviceGetByPCIBusId;
@@ -1699,6 +1702,10 @@ typedef struct hip_api_data_s {
       int peerDeviceId;
       unsigned int flags;
     } hipDeviceEnablePeerAccess;
+    struct {
+      enum hipFlushGPUDirectRDMAWritesTarget target;
+      enum hipFlushGPUDirectRDMAWritesScope scope;
+    } hipDeviceFlushGPUDirectRDMAWrites;
     struct {
       hipDevice_t* device;
       hipDevice_t device__val;
@@ -4609,6 +4616,11 @@ typedef struct hip_api_data_s {
 #define INIT_hipDeviceEnablePeerAccess_CB_ARGS_DATA(cb_data) { \
   cb_data.args.hipDeviceEnablePeerAccess.peerDeviceId = (int)peerDeviceId; \
   cb_data.args.hipDeviceEnablePeerAccess.flags = (unsigned int)flags; \
+};
+// hipDeviceFlushGPUDirectRDMAWrites[('hipFlushGPUDirectRDMAWritesTarget', 'target'), ('hipFlushGPUDirectRDMAWritesScope', 'scope')]
+#define INIT_hipDeviceFlushGPUDirectRDMAWrites_CB_ARGS_DATA(cb_data) { \
+  cb_data.args.hipDeviceFlushGPUDirectRDMAWrites.target = (hipFlushGPUDirectRDMAWritesTarget)target; \
+  cb_data.args.hipDeviceFlushGPUDirectRDMAWrites.scope = (hipFlushGPUDirectRDMAWritesScope)scope; \
 };
 // hipDeviceGet[('hipDevice_t*', 'device'), ('int', 'ordinal')]
 #define INIT_hipDeviceGet_CB_ARGS_DATA(cb_data) { \
@@ -7543,6 +7555,9 @@ static inline void hipApiArgsInit(hip_api_id_t id, hip_api_data_t* data) {
 // hipDeviceEnablePeerAccess[('int', 'peerDeviceId'), ('unsigned int', 'flags')]
     case HIP_API_ID_hipDeviceEnablePeerAccess:
       break;
+// hipDeviceFlushGPUDirectRDMAWrites[('hipFlushGPUDirectRDMAWritesTarget', 'target'), ('hipFlushGPUDirectRDMAWritesScope', 'scope')]
+    case HIP_API_ID_hipDeviceFlushGPUDirectRDMAWrites:
+      break;
 // hipDeviceGet[('hipDevice_t*', 'device'), ('int', 'ordinal')]
     case HIP_API_ID_hipDeviceGet:
       if (data->args.hipDeviceGet.device) data->args.hipDeviceGet.device__val = *(data->args.hipDeviceGet.device);
@@ -9562,6 +9577,12 @@ static inline const char* hipApiString(hip_api_id_t id, const hip_api_data_t* da
       oss << "hipDeviceEnablePeerAccess(";
       oss << "peerDeviceId="; roctracer::hip_support::detail::operator<<(oss, data->args.hipDeviceEnablePeerAccess.peerDeviceId);
       oss << ", flags="; roctracer::hip_support::detail::operator<<(oss, data->args.hipDeviceEnablePeerAccess.flags);
+      oss << ")";
+    break;
+    case HIP_API_ID_hipDeviceFlushGPUDirectRDMAWrites:
+      oss << "hipDeviceFlushGPUDirectRDMAWrites(";
+      oss << "target="; roctracer::hip_support::detail::operator<<(oss, data->args.hipDeviceFlushGPUDirectRDMAWrites.target);
+      oss << ", scope="; roctracer::hip_support::detail::operator<<(oss, data->args.hipDeviceFlushGPUDirectRDMAWrites.scope);
       oss << ")";
     break;
     case HIP_API_ID_hipDeviceGet: