+++ projects/rocprofiler-sdk/source/include/rocprofiler-sdk/hip/api_args.h
@@ -3640,6 +3640,13 @@ typedef union rocprofiler_hip_api_args_t
         unsigned int flags;
     } hipInitDevice;
 #endif
+#if HIP_RUNTIME_API_TABLE_STEP_VERSION >= 34
+    struct
+    {
+        enum hipFlushGPUDirectRDMAWritesTarget target;
+        enum hipFlushGPUDirectRDMAWritesScope  scope;
+    } hipDeviceFlushGPUDirectRDMAWrites;
+#endif
 } rocprofiler_hip_api_args_t;
 
 ROCPROFILER_EXTERN_C_FINI